[win] Improvements by dshoreman v1.5
[win] Expand search to include class and title

Greatly improves usefulness in cases where many windows are open of the
same application, for instance multiple Terminal or browser instances,
where previously you could only search by the program name.

Especially helps in the case of Firefox where the instance is not
"Firefox" but instead "Navigator", meaning prior to this change you
would be forced to search for Navigator to get any results.

By including the window title in matches, we can now also filter by
webpage (for browsers) or document titles (for editors). Previously
where "Alacritty" could max out the results, we can instead type "vim"
to get far fewer results with a higher likelihood of a good match.

[win] Fix icon for Firefox windows

Firefox has a `WM_CLASS` of "Navigator.Firefox", so searching for an
icon based on the instance part (Navigator) doesn't find anything.

This updates it to also search for the lower-case version of the actual
window class which correctly shows the Firefox icon in its results.

[win] Fix windows with spaces in WM_CLASS

This commit fixes an issue where spaces in a windows's `WM_CLASS`
attribute would be split incorrectly causing output to break.

`Navigator.Firefox Developer Edition` would previously have its wm_class
set to `Navigator.Firefox`, with `host` set to "Developer" and `wm_name`
left set to "Edition <actual_wm_name>".

By replacing spaces with dashes, we have the added benefit that the
correct icon is now displayed for the developer edition of Firefox where
before it would show the main Firefox icon.