[youtube] Fix 'simpleText' key error
Fixes an issue that was preventing `yt foo` style commands from working.

While channel results still use a `simpleText` key for the title, video
results now have a `runs` list that contains the title as a `text` key
in the first object. There's also an accessibility object that contains
all data about the video as a string, but we don't need that.