[youtube] Properly load text from `runs` list
While video titles seem to be stored as one string, the video count text
is stored with the number in the first item, and the text in the second.

This adds a small helper function that grabs the `text` value from all
the items in `runs` and concatenates them, falling back to simpleText
where the `runs` key isn't defined so we can use it for titles too.

Also imports the `critical` function from `albertv0` since it's needed
to catch any exceptions that get thrown (eg for shelf/movie renderers).