Protect window_switcher from split errors
This commit stops window_switcher from erroring when the wm_class has
more or less than a single `.` character.  For example, here are two
problematic lines of outputs from `wmctrl -l -x` where the `wm_class` is
`N/A` or `rufus-3.17.exe.rufus.3.17.exe` (from Wine):

    0x01a00078  1 N/A                            remus Ozone X11
    0x03e00003  2 rufus-3.17.exe.rufus-3.17.exe  remus Rufus 3.17.1846

If there are no dots, the entry is skipped and not matched.  If there is
more than one dot, the part before the first dot is used for the display
name (e.g. "rufus-3").